+++ CHANGELOG.md
@@ -63,6 +63,7 @@
   ([GH-1058](https://github.com/NVIDIA/warp/issues/1058)).
 - Fix direct initialization and slice assignment of vector, quaternion, and transform types from PyTorch tensors
   ([GH-1121](https://github.com/NVIDIA/warp/issues/1121)).
+- Fix a potential OverflowError when slicing arrays with np.int32 strides ([GH-1120](https://github.com/NVIDIA/warp/issues/1120)).
 
 ## [1.10.1] - 2025-12-01